Removed various casts.

+++ examples/src/asp_toplass/OrcaRandom.java
@@ -46,7 +46,7 @@ public class OrcaRandom {
         current[2] = (fac[2] * current[2] + term[2]) % mod[2];
         ix = (table_size * current[2]) / mod[2];
         r = table[ix];
-        table[ix] = ((double) current[0] + (double) current[1] * inv_mod[1])
+        table[ix] = (current[0] + current[1] * inv_mod[1])
                 * inv_mod[0];
         return r;
     }
@@ -66,7 +66,7 @@ public class OrcaRandom {
         }
 
         for (int i = 0; i < 3; i++) {
-            inv_mod[i] = 1.0 / (double) mod[i];
+            inv_mod[i] = 1.0 / mod[i];
         }
 
         current[0] = (term[0] + seed) % mod[0];
@@ -78,7 +78,7 @@ public class OrcaRandom {
         for (int i = 0; i < table_size; i++) {
             current[0] = (fac[0] * current[0] + term[0]) % mod[0];
             current[1] = (fac[1] * current[1] + term[1]) % mod[1];
-            table[i] = ((double) current[0] + (double) current[1] * inv_mod[1])
+            table[i] = (current[0] + current[1] * inv_mod[1])
                     * inv_mod[0];
         }
     }
