+++ examples/src/radix/Radix.java
@@ -114,7 +114,7 @@ public class Radix {
         double d2 = Math.log(2);
         double t = d1 / d2;
         int tmp = (int) (t);
-        if ((double) tmp != t) {
+        if (tmp != t) {
             System.out.println(" Radix must be a power of 2");
             System.exit(1);
         }
